backend/requirements.txt
Flask>=2.3
Flask-SQLAlchemy>=3.1
python-dotenv>=1.0
requests>=2.31
youtube-transcript-api>=0.6
yt-dlp>=2024.8.6
